fix typo in Embedded_C/A1/EX2 line 17

+++ 02_Embedded_C/Assignment_1/A1_EX1_Toggle_Led/Inc/hw_access.h
@@ -14,7 +14,7 @@
 #define RO					const volatile
 
 #define IO_RW_REG8(x)		(*((RW uint8 *)(x)))
-#define IO_RO_REG8(x)		(*((RO uint88 *)(x)))
+#define IO_RO_REG8(x)		(*((RO uint8 *)(x)))
 
 #define IO_RW_REG16(x)		(*((RW uint16 *)(x)))
 #define IO_RO_REG16(x)		(*((RO uint16 *)(x)))
